csci5410-summer-23-sdp34-main/Notification_module/TeamUpdate/requirements.txt
# Function dependencies, for example:
# package>=version
boto3
google-cloud-firestore